apiGateway/requirements.txt
Flask==2.1.0
Flask-Compress==1.8.0
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.1.1
requests==2.27.1
PyJWT==2.3.0